requirement.txt
streamlit==1.10.0
PyMuPDF==1.19.6
pandas==1.3.5
plotly==5.3.1
google-generativeai==0.1.0  
datasets==1.17.0
